done movie, create view for user, fix some error

+++ Areas/Admin/Controllers/SongController.cs
@@ -1,4 +1,5 @@
-﻿using DVDShops.Models;
+﻿using DVDShops.Helpers;
+using DVDShops.Models;
 using DVDShops.Services.Albums;
 using DVDShops.Services.AlbumsGenres;
 using DVDShops.Services.Artists;
@@ -9,6 +10,7 @@ using DVDShops.Services.SongsGenres;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.IdentityModel.Tokens;
 using System.Diagnostics;
+using System.Text.RegularExpressions;
 using static System.Net.WebRequestMethods;
 
 namespace DVDShops.Areas.Admin.Controllers
@@ -56,11 +58,17 @@ namespace DVDShops.Areas.Admin.Controllers
 
         [Route("addSong")]
         [HttpPost]
-        public IActionResult AddSong(Song song, IFormFile songFile, List<string> genres)
-        {            
-            if (string.IsNullOrWhiteSpace(song.SongName) || string.IsNullOrWhiteSpace(song.SongIntroduction))
+        public IActionResult AddSong(Song song, List<string> genres)
+        {
+            if (string.IsNullOrWhiteSpace(song.SongName) || string.IsNullOrWhiteSpace(song.SongIntroduction) || string.IsNullOrWhiteSpace(song.DownloadLink))
             {
                 SetTempData(false, "Create Song Failed!", "Some Input Field Is White Space Only!");
+                return View("SongAdd", song);
+            }
+
+            if (!CheckRegex.CheckLink(song.DownloadLink))
+            {
+                SetTempData(false, "Create Song Failed!", "Url is Invalid!");
                 return View("SongAdd", song);
             }
 
@@ -79,27 +87,6 @@ namespace DVDShops.Areas.Admin.Controllers
             if (song.ArtistId < 0)
             {
                 SetTempData(false, "Create Song Failed!", "Please Choose an Artist!");
-                return View("SongAdd", song);
-            }
-
-            if (songFile != null && songFile.Length > 0)
-            {
-                string fileExtension = Path.GetExtension(songFile.FileName);
-                if (fileExtension != ".mp3" && fileExtension != ".wav" && fileExtension != ".ogg")
-                {
-                    SetTempData(false, "Create Song Failed!", "Please Choose Correct File Extension!");
-                    return View("SongAdd", song);
-                }
-                song.DownloadLink = Guid.NewGuid().ToString() + "_" + songFile.FileName;
-                var path = Path.Combine(env.WebRootPath, "/download/song", song.DownloadLink);
-                using (var stream = new FileStream(path, FileMode.Create))
-                {
-                    songFile.CopyTo(stream);
-                }
-            }
-            else
-            {
-                SetTempData(false, "Create Song Failed!", "Please Upload The Song!");
                 return View("SongAdd", song);
             }
 
@@ -132,6 +119,11 @@ namespace DVDShops.Areas.Admin.Controllers
         public IActionResult DeleteSong(int songId)
         {
             var song = songService.GetById(songId);
+            if (song.AlbumsSongs.Any(als => als.SongId == song.SongId))
+            {
+                SetTempData(false, "Delete Song Failed!", "This Song Cannot be Deleted This Way!");
+                return RedirectToAction("view");
+            }
 
             if (songService.Delete(songId))
             {
@@ -154,8 +146,27 @@ namespace DVDShops.Areas.Admin.Controllers
 
         [Route("editSong")]
         [HttpPost]
-        public IActionResult EditSong(Song song, IFormFile songFile, List<string> genres)
+        public IActionResult EditSong(Song song, List<string> genres, string newLink)
         {
+            if (!string.IsNullOrWhiteSpace(newLink))
+            {
+                if (newLink.Trim() == song.DownloadLink)
+                {
+                    SetTempData(false, "Update Song Failed!", "New Download Link Cannot be The Same As Current Link!");
+                    return View("SongEdit", song);
+                }
+
+                if (!CheckRegex.CheckLink(newLink))
+                {
+                    SetTempData(false, "Update Song Failed!", "Download Link Invalid!");
+                    return View("SongEdit", song);
+                }
+                if (!string.IsNullOrEmpty(newLink))
+                {
+                    song.DownloadLink = newLink;
+                }
+            }
+
             if (string.IsNullOrWhiteSpace(song.SongName) || string.IsNullOrWhiteSpace(song.SongIntroduction))
             {
                 SetTempData(false, "Update Song Failed!", "Some Input Field Is White Space Only!");
@@ -178,29 +189,6 @@ namespace DVDShops.Areas.Admin.Controllers
             {
                 SetTempData(false, "Update Song Failed!", "Please Choose an Artist!");
                 return View("SongEdit", song);
-            }
-
-            if (songFile != null && songFile.Length > 0)
-            {
-                string fileExtension = Path.GetExtension(songFile.FileName);
-                if (fileExtension != ".mp3" && fileExtension != ".wav" && fileExtension != ".ogg")
-                {
-                    SetTempData(false, "Update Song Failed!", "Please Choose Correct File Extension!");
-                    return View("SongEdit", song);
-                }
-
-                var oldFile = Path.Combine(env.WebRootPath, "/download/song", song.DownloadLink);
-                if (System.IO.File.Exists(oldFile))
-                {
-                    System.IO.File.Delete(oldFile);
-                }
-
-                song.DownloadLink = Guid.NewGuid().ToString() + "_" + songFile.FileName;
-                var path = Path.Combine(env.WebRootPath, "/download/song", song.DownloadLink);
-                using (var stream = new FileStream(path, FileMode.Create))
-                {
-                    songFile.CopyTo(stream);
-                }
             }
 
             var result = songService.Update(song);
